Basic files
Adding a couple of basic files for Project 3. BikePart has the methods to create objects of type bikePart and to create the ArrayList. WarehouseInventory has an iterator that has not been implemented yet- still working around with it.  WarehousePart has two comparators- one to sort by name and one to sort by number.